重构代码，删除obs_utils
